module_5/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
graphviz==0.20.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydeps==3.0.1
stdlib-list==0.11.1
Werkzeug==3.1.3
pylint==3.1.1